Fix the performance issue of the Derived Lines drawing
An performance issue was reported that when many ConstraintSets Derived Lines need to be drawn, the Android Studio UI would freeze for serveral minutes.

To draw the derived lines, we need to first determine the level (PathYOffset) of each line to minimize the number of line crossings. However, the existing approach requires O(3^n) time complexity to calculate the levels, where n = the number of derived lines. Therefoer, the performance can get worst rapidly when n grows.

In this CL, an alternative approach is implemented to improve the time complexity to approximate O(n^3). Although this new approach can not guarantee the best result, it still offers a relatively good result with a much lower time complexity.

Bug: 197401536
Test: OverviewPanelTest + Manually tested
Change-Id: I87966ae2efcaf773fd6d6eee2697eee38573681e