Autoupdate of 2 submodules
*tools/adt/idea updated to cf3fd980a8dec27ac43031232ec5179a1ab13279
Update Wizard to support Material 3 Compose Actitivty

This moves from using Category to hide the templates behind a flag to
using Constraints.

Bug: 200005771
Test: Existing tests pass
Change-Id: I37e40ea6e37549cd3d7dd8c6bff8ce930b2d0ef8

*tools/base updated to 017c779e477802c7affe6e95b19fadfb03657148
Update Wizard to support Material 3 Compose Actitivty

This moves from using Category to hide the templates behind a flag to
using Constraints instead.

The template test for Compose is disabled since the only artifacts
available are currently the snapshots. The test verifies that the
templates are not enabled accidentally for users.

Bug: 200005771
Test: Existing tests pass
Change-Id: Icdf38ff6d77b3b367d517eea0e73f4aecc8a71e0