Autoupdate of 2 submodules
*tools/analytics-library updated to db473fe1449fb832e0b774763e8b38b9e5499640
Expose whether the UsageTracker is initialized

This is neccessary for lint to make sure it initializes
the usage tracker if it hasn't already been done. When
running in process with AGP, AGP will have initialized it,
but when running as a separate command, it has not.
Using just AnalyticsSettings.initialized as a proxy didn't
work well since for example during unit testing the
analytics tests will initialize the settings but close
the usage tracker when done.

Bug: 194525628
Test: N/A (just making a getter public; functionality tested in lint)
Change-Id: Ie5cf067c7918bc27764d95790075fa9fe0b3f157

*tools/base updated to 300ae2042a7ba261c4729dfb151b62dab6ee02b9
194525628: "call to UsageTracker before initialization"

This CL fixes analytics computations for lint when
running outside of the IDE.

Test: LintBatchAnalyticsTest and ChildCountDetectorTest
Fixes: 194525628
Change-Id: I3fbabffa6d2b8e605fb44b7ad4dddefa5fb9b1cc