Autoupdate of 2 submodules
*tools/base updated to 519bce8f579bc0f4de10dae9af2c3ae0a906a8e5
Address perf Dolphin perf regression

This CL tries to address the perf regression in the referenced bug.
The change I believe caused the issue is in CxxRepublishBuilder.
Before ag/15882890 CxxRepublishBuilder did not republish .a libraries
and afterward it did. The reason is subtle: the prior code assumed
each republished file was in the root of the SO folder. The .a libs
for Dolphin were not in the folder so they weren't republished. I
believe it is correct to only republish SO files so I've changed
the code to work that way.

I also made a change to respect the original subfolder arrangement
in SO folder. So, for example:

  [SO folder]/a/b/c/lib.so => [Republish folder]/a/b/c/lib.so

Whereas before it was:

  [SO folder]/a/b/c/lib.so => [Republish folder]/lib.so

That is, the subfolder information was lost. For most projects,
this won't matter but it could have been a problem in cases where
the same lib name was present in multiple sub-folders.

Lastly, this CL, now enforces that only libraries from [SO folder]
are republished. This change means that Dolphin project no longer
republishes test executables since these are built to a folder
other than [SO folder]. The reason I did that is because these files
can also collide if they have the same name as a file in [SO folder].
If we find we really do need to republish files like this then we
will need to invent a scheme to avoid collision. However, I don't
believe that will be necessary.

As part of this CL, I added an integration test that runs the actual
Dolphin project outside of perfgate. I use this to monitor for
regressions in terms of counts of expensive operations such as copying
files and executing processes. I hope this can catch some regressions
in PSQ rather than waiting for a perfgate run.

Bug: 201546326
Test: Added DolphinProbe test which would have failed if it had been
  in place before this regression.

Change-Id: I1c613ec80c7182b19342b2fb4578cfef710110b4

*prebuilts/studio/buildbenchmarks updated to 2f7c4ad8971b4206c0a7e5056d7de8fcd9c13907
Add Dolphin setups for CMake 3.18.1 and 3.10.2

The default for perfgate continues to be 3.10.2 (that's set up by setup.diff).
This CL establishes explicitly named setup-cmake.3.10.2.diff and
setup-cmake.3.18.1.diff to use in the DolphinProbe test project.

Bug: 201546326
Test: Supports DolphinProbe test.
Change-Id: I48dd411611f18a4d3e8ef70161164a1925a75729