Autoupdate of 2 submodules
*tools/adt/idea updated to 3cabee126a80c1494468cc426948c2d3b44c2bcd
[MPSS] Use main modules in DataBinding tests

When we switch to module per source set the top level module for each
Gradle project will no longer have sources. This change ensures that the
main modules are used where sources are required.

Bug: N/A
Test: N/A
Change-Id: Ib5ea3245aaf8c60298fd6b6cf93e87e0a4ef2bd9

*tools/vendor/google updated to f2bf863d4bba82b82eb850c50a27cc3d34cecfc0
[MPSS] Remove module listings from source providers snapshot

We do not need these tests to rely on the number and name of the modules
they should just verify the data in the Android facets, the names and
structure of the project are verified via the other types of snapshot
tests.

Bug: N/A
Test: N/A
Change-Id: I9bf91674b00cd5404402b1d0c8833de01ee316c9