Temporarily disable ATF in visual lint
I think turning off obtain char loc feature reduces the render time
significantly enough that this CL is probably not necessary.

However, in the case that we still wish to disable it, I am ok with
the decision (for visual lint). This is the CL to do so.

We can abandon it if we feel it's not needed, but creating an CL for
an option.

Bug: N/A
Test: Manually tested
Change-Id: I19cdc9fd25442e8e606639d81b33e177084b5cbe