Autoupdate of 2 submodules
*tools/adt/idea updated to 3c1b5078c6db3e546da046e1d775b04e38269a6a
[Hybrid build] Add studio checks to tools/adt/idea

To allow editing lint all in one place and unblock the removal of lint
from the AGP gradle project.

This change also removes lint-gradle from the modules.xml, which was
accidentally introduced in 30772be8df65.

Bug: 143864616
Test: included
Change-Id: I9719dc4ea600ae589efe66880c57673b876a1755

*tools/base updated to 9608830906d1a63837e62a2b31f7f5409a39557a
[Hybrid build] Add studio checks to tools/adt/idea

To allow editing lint all in one place and unblock the removal of lint
from the AGP gradle project.

Bug: 143864616
Test: included
Change-Id: I086d14bcd5e618a68bf56f0b6d03ffd9648baaa7