Simplify the way we choose which artifacts to build
for a given run configuration.

Android Studio supports a hardcoded list of known run configuration
types in addition to its own run configurations.

This change makes it possible for run configurations intended to work
in Android Studio to specify which artifacts to build via the existing
`PreferGradleMake` interface instead of hard-coding their names in the
Gradle project system support code.

The list of other supported run configuration types remains to be
hardcoded.

Bug: 150462790
Test: existing
Change-Id: I5c417f5f3370dd8fd20a966510ee3d56e5f5fc04