Autoupdate of 2 submodules
*prebuilts/r8 updated to e2795827e12ae864c7105ae5ca5aac402d6e128f
Update D8 R8 to 3.2.10-dev

This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.2.10-dev
Test: ./gradlew check
Bug:

Change-Id: I7b6dccf51572c8fc48245fb3c9f4b26f2a2726b7

*tools/base updated to 122c8639b355bfbaee8a2198bf1c0c579e0207e9
Update sorting of keep file output in tests.

This has been changed as part of making the sorting deterministic in R8

Also update gmaven.txt after we added the api_database to r8.

Bug: 202697611

Test: CoreLibraryDesugarTest

Change-Id: Ice944a3746bc02487caaa69eda4215d7e6f7cc19