Autoupdate of 3 submodules
*prebuilts/tools updated to 378a93166275769cf55d5ca5bca6887cbb262a99
Add ui-test-manifest artifact for Compose

This artifact is now part of the default Compose templates and required
for testing the templates.

Bug: 202839436
Test: Added for TemplateTest
Change-Id: Ib9221da9a775bd69a163fdfa51de68726c0f4d31

*tools/adt/idea updated to 99e17fa9c32c9e513ed05918bf7e616589ddbc18
Add ui-test-manifest artifact for Compose template

This artifact is now part of the default Compose templates and required
for testing the templates.

Bug: 202839436
Test: Added for TemplateTest
Change-Id: I811cccd565accf0ac6ed3892e0b26898666ef730

*tools/base updated to 31efb399a846a1acbd34d33ca9e242786cf97df7
Add compose:ui:ui-test-manifest to Compose template

For tests to be able to use `createComposeRule()`, the manifest artifact
needs to be added as a debugImplementation dependency. This will add
ComponentActivity to the manifest.

Bug: 202839436
Test: TemplateTest
Change-Id: I1b116c84a26ba7f0a35a412bfd24b07f484fe134