Autoupdate of 2 submodules
*tools/adt/idea updated to 90569c3540481c18f4e744fe9eaa6ecf47fbe34e
Move AS configuration from platform/tools/idea

The configuration code is moved into androidstudio.xml, maintaining
the condition that this code only runs in Android Studio (and not
in the Android plugin running in IDEA).

This makes it easier to change the code without having to rebuild
the platform.

Bug: 200037669
Test: existing tests (e.g. AndroidStudioUpdateStrategyTest)
      along with some manual inspection in the debugger to ensure
      all of this code still runs
Change-Id: I221ac46c3dadbcb6a8bde07eb2dbd78a91771b2d

*prebuilts/studio/intellij-sdk updated to c779408bb5e6a534249165eb12ba3f266a0f4c04
Update AI-212 after Change I52ca06406

Change-Id: I8227a8ff0179e165b681250488cbb9f352e9649b