Revert "Don't select configuration if deploying preview from toolbar"
This reverts commit 3a86e8a9156a1513313caceab910d641fa19ffe3.

Reason for revert: this change was done to address a UX related issue,
but it happened to become inconsistent with the IntelliJ guidelines.
Also, another reason for this revert is to avoid having different
behaviours in the buttons used to deploy a preview to a device (the
one in the gutter and the one in the toolbar).

Fixes: 184687415
Test: manually, by running a project and checking the behaviour of the
buttons.

Change-Id: I48a2977dfc98a90fe1ad4ab8f549e6c588f27bcd