Fix how [more] array elements are populated
When there are more than 50 elements in an array/struct,
an artificial node is placed at the 50th child in the
variable tree. Double clicking this child node is supposed
to render the remaining 50 elements.

There was a change in CIDR that we received with Arctic Fox
that started passing "nextIndex" as a parameter, which
conflicted with a very old local change we had
(I don't know what purpose that one serves).

This CL fixes the conflict, and uses the correct method for
populating the rest of the array/struct.

Before:
  1. https://screenshot.googleplex.com/BHmD2EktJdrdtKg
  2. https://screenshot.googleplex.com/RqNjJtdbhD9ARRG
After:
  1. https://screenshot.googleplex.com/B2iFaYE5A65ydw4
  2. https://screenshot.googleplex.com/4WuHxtTntxFKfQ8

Note on testing: I tried adding an integration test for this,
but our LLDB testing infra is headless, so it bypasses this
UI related part about adding synthetic [more] node. That's why
this CL has no automated integration tests.

Bug: 202470196
Test: manual
Change-Id: I209a40e569375fc5037ab0edb96c7ef6579bc47c