Autoupdate of 2 submodules
*tools/adt/idea updated to 5a4719ed2ea11a55839be03751e09f27b7226341
Disable logging unit tests execution on the IDE side.

in
https://studio-stats.corp.google.com/#report_id=Testing/Test%20runs%20total
we get unit test execution logged twice : this is because we do log the
test execution on the IDE side, and again on the AGP side in order to
get the test libraries being used.

Logging tests execution on the IDE side is not providing any additional
information than the AGP's, so it makes more sense to log events from
the AGP side only.

Bug: N/A
Test: on AGP side
Change-Id: Iaeeb195e7febd52557e293f411f1ad7e62ebea7c

*tools/base updated to 48a2fccb3d04eca3d471c23c59edc459a9dd3829
Change invocation type when logging unit tests metrics on AGP.

The tests execution that we log refers to execution from Android Studio,
and not Gradle.

Bug: N/A
Test: N/A
Change-Id: If480036171a93db1f1c14d5a0502c0c13246aba6