Autoupdate of 2 submodules
*prebuilts/studio/layoutlib updated to 5bc2836cbd6a936d4c1fe350787650398243c51d
Add layoutlib-extensions as an artifact

This allows to be used from tests.

Bug: 162056408
Test: N/A
Change-Id: Ifca4aea922224205d80b77c7c267dc15e4c289c5

*tools/adt/idea updated to 14ebbcfe218ea2e82164c3b31cf3ebbc37b5d129
Fix regression in Android coroutines initialization

The kotlin coroutines core library will detect that is running in
Android and try to load "AndroidDispatcherFactory" to create the
coroutines dispatchers. By making it fail, we force the
kotlinx.coroutines.internal.FastServiceLoader to fallback to the backup
loading method. This method involves trying to find the right dispatcher
factory by locating the
kotlinx.coroutines.internal.MainDispatcherFactory service and
instantiating it.
We have our own implementation of that service and we declare it as part
of the layoutlib-extensions module that is loaded as an external jar.
The service will load the LayoutlibDispatcherFactory that implements a
dispatcher adapter to run on layoutlib.

Test: Added regression test
Bug: 162056408
Change-Id: I546474c415c5eb26a2f63e217cf94802a7b3c224