Allow PseudoClasses to be found in the plugin class loader
When locating PseudoClasses during the ASM transformation phases,
classes might not be found in the project class path. Classes that were
loading from the plugin, like java.* need to still be resolved to a
PseudoClass to be able to apply transformations correctly.

This change allows the PseudoClassLocatorForLoader to fallback to
loading the classes via ClassLoader.getResource or
ClassLoader.findClass.

Fixes: 196985631
Test: Added regression test
Change-Id: Ie3cef790853e3540dbc5efb6dc92adac78e8631e