Split SyncVariantResult into smaller pieces
Sync executes in different modes and depending on the current mode
different `SyncVariantResult`'s properties are used. This change
begins splitting the structure into more granular, in the future,
immutable structures.

(1) Extract `SyncVariantResultCore` (which is to be renamed in
    upcoming changes) to hold non dependency related data, for now,
    for both v1 and v2 sync modes

(2) Wrap it into `SyncVariantResult` (which is to be renamed in later
    changes) representing basic and dependency data.

Bug: n/a
Test: n/a
Change-Id: Idedae14d3f7160d76d392896bd192c3e28d6f92c