Autoupdate of 3 submodules
*tools/adt/idea updated to a54c21f2a32062e0f4432ade937a1bb810fddf89
Add a test for shortcuts.xml

- Change the test name for App Actions accordingly

Bug: 177019730
Test: This is a test
Change-Id: I2dc63fbf3170e7c5a11615f283a49e6c1c22a5fe

*tools/analytics-library updated to 83c06131c2af168be49664a6adf0a3abfd16ef56
Update studio_stats.proto

Logs-Approval: http://cl/402755368
Bug: 177019730
Test: existing
Change-Id: I0efb57564b6a1babd8d9ca66a66b78b79e92aeaa

*tools/base updated to ff9ad07abbb8a7071c712e86d8caed7fbc244426
Add a menu item for creating shortcuts.xml

Since the App Actions are deprecated, we need to replace the menu to
create an XML for app actions with an XML creating shortcuts.

The menu item for App Actions is not removed at the moment, but only
marked as deprecated.

Bug: 177019730
Test: TemplateTest, manually created the shortcuts from the context menu
Change-Id: Ie1230890001adff8c27e1d15a25d8659bb46bb14