Avoid keeping references to Gradle builder models between sync stages
by converting them to incomplete IDE models as soon as is possible.

This change (1) is a step towards separation of libraries and
dependencies in IDE models and (2) reduces the memory usage of
sync since the IDE models are smaller than the builder models
and, moreover, the builder models are wrapped into proxies.

Bug: 200956871
Test: n/a (existing)
Change-Id: Ia841f06226d6bacfe3957bb4d17e8c23c60262c8