Remove spurious dsl.kotlin dependency from core
The kotlin support in general is required at runtime / test-time, but
not to compile most things, as the interface provided by the gradle-dsl
module should be sufficient.  Modify the testutils module to have
runtime-scope dependencies on the test-runtime needs of dsl clients, and
depend on that testutils module in modules whose tests transitively end
up using the dsl layer.

Bug: none
Test: existing
Change-Id: I38c4b4162a72484f84d9699003cebd56152af395