Autoupdate of 2 submodules
*tools/adt/idea updated to abae698f95fa56ededb29d7d5a34898c6ffd7336
Revert "Update model snapshot for testFixtures project"

This reverts commit 809ce7fc3eab7c9675f69d8f1f6bd56e156d4c45.

Reason for revert: variantAttr is no longer used in v2 model, however
in v1, we still need to select the correct variant in case of
testFixtures module dependency. And since testFixtures is not
supported in v1 models, it's better to set the variantAttr to the
main variant instead of the testFixtures variant so the IDE could
pick up the main variant directly.

Change-Id: I5ffa556909c3918e7428243368c7f49e742144be

*tools/base updated to e31caf3508f3f4b0af508131059c291f90683b88
Revert "Use testFixtures variant name for testFixtures resolved artifacts"

This reverts commit 5a3a1f6186eadd96d92f4582bed4b447b0b8a33e.

Reason for revert: variantAttr is no longer used in v2 model, however
in v1, we still need to select the correct variant in case of
testFixtures module dependency. And since testFixtures is not
supported in v1 models, it's better to set the variantAttr to the
main variant instead of the testFixtures variant so the IDE could
pick up the main variant directly.

Change-Id: I0e2613b76ec2f6076201c8256d3c109923476322