Autoupdate of 3 submodules
*prebuilts/tools updated to 29a34aec069b2272bae062e71911cb8f2a7d6540
Update to Compose 1.1.0-beta01

Remove 1.0.3 version.

Test: Existing tests pass
Bug: 203585777
Change-Id: I4acce078cb28a0b4c25176a50aa3ded7d734008d

*tools/adt/idea updated to e6626773dfc651925c7136870a7aa7b7d9495184
Update to Compose 1.1.0-beta01

This CL temporarily pins the template and designer tests to API 31 until
the update of API 31 is complete and becomes the latest in the tree.

Test: Existing tests pass
Bug: 203646112
Bug: 203585777
Change-Id: Ibbdcb4bbb894c824f68a84ea240c4b4510279d69

*tools/base updated to 71537660bdc8a11cff10582c0c51092fda88f538
Add Compose 1.1.0-beta01

Temporarily allowing Compose related tests to see the
platforms/android-31 rule. This will be reverted once 31 is available
for all tests as "latest".

Test: Existing tests pass
Bug: 203585777
Bug: 203646112
Change-Id: I298eb1f1855590cc3e2bd538d8be4e472a37d627