Display frame token in trace event name
In S, Perfetto appends frame token to the trace event, e.g.
DrawFrame 1234. Currently profiler replaces any number suffix with
"###" to normalize the name for things like color-coding and
occurrence count. However, we still want to display the frame token
in Call Chart and the Events tab to help user associate them with
the frame in Jank Detection.

Fixes: 187425456
Test: updated
Change-Id: I4d9627058f65ce79ed7041e5ba51f8a389318800