Autoupdate of 2 submodules
*tools/adt/idea updated to 4dec9b25a8aa3e956ae8aea35b9f7f3d1cfa960d
Support for optional sourcesets in models.

AGP will now exclude sourcesets that aren't used
by any variant or test components. This means some
of the sourceset will be nullable. This change updates
the snapshots and the model and consuming code to
handle this.

Fixes: 169403306
Test: N/A
Change-Id: I033d6ae9b6f6e415d6649a28e2790a71cd1aa287

*tools/base updated to 6cfe088b73293414a779cf1f0fc0e16434ef64d5
Don't include unused sourcesets.

If a sourceset is not used by a variant or a test
component, do not include the sourcesets in the model.

Including it can be confusing to users since the folder would
be mark as a source but won't actually be used.

Fixes: 169403306
Test: updated existing snapshots + new test validating change on disabled variants.
Change-Id: Idcd41f19ef050c313037bde72711495dbb1df4dd