Autoupdate of 2 submodules
*tools/adt/idea updated to a6294320bba93caeec9c61d5e660fc418f43e5ea
Use ProjectInfo.istestFixtures

Previously we were searching in the capabilities for a testFixtures
capability to indicate whether a dependency is on the testFixtures
component of a module. This change moves to using the isTestFixtures
value directly.

Bug: 139438142
Fixes: 202838863
Test: existing
Change-Id: I52a8dcaa68ac7c656374ddc7442d76fa61386b7d

*tools/base updated to e351c360a75223d97eec4480f0f18ff334251f04
Add a boolean value in ComponentInfo to mark a testFixtures component

Add a boolean value in ComponentInfo to indicate if a library or a
module is a testFixtures component. This is needed to eliminate the need
for the IDE to check the capabilities for a testFixtures capability and
move the checking logic to AGP where we know more information about the
component.

Bug: 139438142
Bug: 202838863
Test: model snapshot tests
Change-Id: I3d09e457a27951121c32ad7a85bad41b7a75c906