Modified revert of "Add couple DexCache read-barriers."
Remove the DexCache read-barriers again (the revert) since
there is measurable performance difference in app startup (~1%).
It turns out they are on very hot path in the interpreter.

The lazy array initialization made the DexCache fields mutable.
However, instead of the read-barriers, we can just ensure we get
the right to-space object on the array-initialization slow path.

Test: test.py -b -r --host
Change-Id: I3f0123141e9a7875a26f088bff68d50641b66883