Autoupdate of 2 submodules
*tools/adt/idea updated to 9f23feea8d5f0e428300dd6657859b94fc9f6018
Avoid Triggering OnItemChange Events on Filter Combo

Rather than removing SelectedProcessFilter and Firebase filter from the combo and then adding new ones with a new ClientData, add a setClient() method to the filter and change the filter in place.

This prevents a OnItemChange event from triggering which causes b/195171994

Fixes: 195171994
Test: Added
Change-Id: Ib4153a40ec4d83a7ef258dce759fb0ad6751028d

*tools/vendor/google updated to 3292dfea56cbc0e81abb22ace67020b284c6798c
Avoid Triggering OnItemChange Events on Filter Combo

Rather than removing SelectedProcessFilter and Firebase filter from the combo and then adding new ones with a new ClientData, add a setClient() method to the filter and change the filter in place.

This prevents a OnItemChange event from triggering which causes b/195171994

Fixes: 195171994
Test: Added
Change-Id: Iadfd4f4ae5b993b5b3cb5e9f36a88bad6ffa7384