Autoupdate of 2 submodules
*tools/adt/idea updated to 44221c8aa5a1d6b0f3dc7e97ad76d80867d2bb99
Update model v2 converter

to match the changed LintOptions properties.

Bug: N/A
Test: existing tests
Change-Id: Id56ceb8d34b6ef163d661e9c2e107d89652a1b34

*tools/base updated to 7ce594de7e4f0d55900e3e077fb74584044b5a9f
Clean up lint options in v2 model

Adds the missing fatal, warning, error and informational
sets.

Changes the name of the properties to match the new names
in the DSL interface.

Removes the severityOverride map as it can be computed
on the IDE side from the list of enable/disable/fatal/warning/...

Bug: N/A
Test: updated the snapshots
Change-Id: I06c48aa04bcaded57637d15b35c7b754d223431b