Calculate project path from Gradle model
This removes a couple of hardcoded path in the Live Edit preview
compiler and moves it to rely on the Gradle model paths.
Also, we remove an extra check that would make the ModuleClassLoader to
not correctly update when the overlay updates the first time.

Use the holder module as module to index the overlays.

Bug: 199384558
Test: Added rendering regression test
Change-Id: I2fa6ae121da6b83fe2a8d484485c5986ac8c1fc8