Autoupdate of 2 submodules
*prebuilts/r8 updated to 8d0dc521d1cabc5f4a897f3b1ab6ac25affe6b60
Update D8 R8 to 3.2.21-dev

This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.2.21-dev
Test: ./gradlew check
Bug:

Change-Id: Id7a9e1db7814d583f7788c309574f44112bd4e9a

*tools/base updated to 42af7d7acd34affcb48e46b25a6bf46dc8f639a2
Fix and disable debug tests

Bug: 205264185
Test: existing

Change-Id: Ibe1e1b4a67314302d6a5070d76ce7022d7a24ec1