Autoupdate of 2 submodules
*prebuilts/r8 updated to 4a990072bddad7094dca108b0c9a72040bb4c547
Update D8 R8 to 3.2.23-dev

This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.2.23-dev
Test: ./gradlew check
Bug:

Change-Id: If3cc12c507aa93d62ef03f115cbd20af42f810c8

*tools/base updated to ee912ba79dfeb270dcd80da1a1ba0eebdf7cc964
Fix test assuming R8 don't dead code eliminate calls to support library.

The reason for this is that even if you run with don't optimize we still
do dead code elimination (because it is not visible if it is there or not)
This now includes synthezising a assume values rule on
Build.Version.SDK_INT (which we already did when optimize was true)

So code like:
void foo() {
  if (Build.Version.SDK_INT < 14) {
    SupportLibA.foobar();
  } else {
    SupportLibB.foobar();
  }
}
will result in:
void foo() {
  SupportLibB.foobar();
}
if we are compiling with min-api 18 (which we are in this test)

Bug: 206059290
Test: //tools/base/build-system/integration-test/application:coupled_tests
Change-Id: I41008c4c6dd72e263e4463c54ab46c1fa1cb975b