[Sync] Use project as parent disposable
Usually, using the project as the parent disposable for a listener is
not a good idea, as it would not dispose any dependents on plugin
unload, leading to leaks.  In our case, though, the service naturally
has the same lifetime as the project as a whole, and this is not plugin
code that is likely to be unloaded from Studio in any kind of normal
circumstance, so depending on the project disposable is not wrong.

Project service instances are meant to be disposed when their project
is; however, there seems to be a race condition between this disposal
and the test suite leak checker, leading to flaky leak detection.

Bug: 206328352
Test: maybe less flaky ConfigureAndroidModuleStepTest?
Change-Id: Iaafe77da7823ff8442874ac32fc1f61b9fc380bc