Autoupdate of 4 submodules
*tools/data-binding updated to 6d5a0b00225a54c53188ec4c6aa09741c9d0bdb8
changed version of secrets-gradle-plugin to 2.0.0

older versions contained calls to AGP on APIs that were never
stable and got removed.

Test: Existing.
Bug: N/A
Change-Id: I9c9c866007239619c266c8d1e16529cdfa87f06f

*prebuilts/tools updated to bb8fc3d2206138828ecb511120945582f75ee89e
remove secrets-gradle-plugin:1.3.0 binary

plugin was updated to 2.0.0

Test: Existing
Bug: N/A
Change-Id: I062abc0cd1354519e8d7302f34e05aae1e784d31

*tools/adt/idea updated to 2962cb15da188eb3d10303a103cdb57e47dbd700
update secrets-gradle-plugin usages to 2.0.0

Update was necessary as version 1.3.0 referenced deprecated APIs
we want to remove in Chipmunk.

Test: Existing
Bug: N/A
Change-Id: I2ff25d3509f070132c45d8788e5650cb6f036d48

*tools/base updated to 3473bc657a88c8413f4c9d4f6ecb8794c32317c7
remove secrets-gradle-plugin version 1.3.0

version 1.3.0 depends on a deprecated AGP API that we want to
remove. Upgrade to 2.0.0 which has a fixed usage.

Test: Existing
Bug: N/A
Change-Id: Ie5955eb4cc88bec805b875bf4626501b5df2d068