[IconPickerDialog] Fix bundled icons not loading
Bundled icons would not load due to a change in the classloader which we
relied to locate files in the bundled jar file.

We now just avoid searching for files and try to load icons directly
based on the accompanying metadata.

On the upside, this means that icons will load much faster, specially
when loading bundled icons, on the downside, this requires that we are
always consistent when naming icons, so icons that do not follow the
expected format will not be loaded.

Fortunately, we were already consistent with the names and the change
also makes sure that the name stays consistent when loading and
downloading icons.

Bug: 202494706
Test: manual from presubmit build, updated unit tests
Change-Id: I46defcc5998e259b2f1767e92ba054e51832c6bc