Autoupdate of 2 submodules
*tools/adt/idea updated to 2461cd8cc793ed09c00f00205976eeca7abafe26
Add tests to AndroidVirtualDeviceTest

AndroidVirtualDevice uses DetailsType.getSysImgPath to
determine the path of a system image based on the latest
AndroidVersion. The implementation of this was changed to
support extensions.

Fixes: 206624415
Test: AndroidVirtualDeviceTest
Change-Id: I7fdb7fd1fc3f9c03cff157b3f4280ce747c5564c

*tools/base updated to 8a0eec1054df2afbacd28ac2a15b3e96adbe8e11
Fix loading SDK with extensions during build

SdkDirectLoadingStrategyTest does not parse the package.xml
to determine the correct SDK to load, instead making a guess
at the path depending on AndroidVersion. This meant it needs
to be updated to include extension information.

Fixes: 206624415
Test: DetailsTypePathsTest SdkDirectLoadingStrategyTest
Change-Id: I45a46717616498aeac155c3f064d8cb8e08fa359