Autoupdate of 2 submodules
*tools/adt/idea updated to c24ca9ca4016f197882117cd611e8424b2b97cc5
Add an identifier property to lint model dependencies

Currently, the artifact address value is being used as a unique
identifier, which is not correct in the case of testFixtures dependency
as the testFixtures and the main artifacts will have the same maven
coordinates.

This change adds an identifier property that properly uniquely
identifies the dependency. For libraries, we get the unique identifier
from the IdeArtifactLibrary name, which is generated uniquely for each
library. And for modules, the identifier is the combination of the
sourceset and the project path.

Test: lint model snapshot tests
Bug: 139438142
Change-Id: Ice53b844fecb81f34a2c305d470385c7431665b9

*tools/base updated to 28c6943c159625ed0a67ac9a03d84a31f1984868
Add an identifier property to lint model dependencies

Currently, the artifact address value is being used as a unique
identifier, which is not correct in the case of testFixtures dependency
as the testFixtures and the main artifacts will have the same maven
coordinates.

This change adds an identifier property that properly uniquely
identifies the dependency.

Test: lint model snapshot tests
Bug: 139438142
Change-Id: Ied4f587eb2f955eb1d4fae85cffc8fa5955b8b80