Use AdbFileProvider in Game-Tools device explorer
ag/15727408 modified device explorer so that it reads the
location of adb.exe from the project using AdbFileProvider.

AFAIU, AdbFileProviderInitializer registers an AdbFileProvider
when a project is opened, and puts this value:
AndroidSdkUtils.getAdb(project) into the provider.

In Game Tools, we create an empty placeholder project, but
we don't open it, so we never get an AdbFileProvider registered.

This CL manually registers an AdbFileProvider for the project.

Bug: 207689600
Test: Run GameTools with: --app-window DEVICE_FILE_EXPLORER
Change-Id: I380ecf09e5877421a9af7ac0c224f131c49d7e49