[Drawable picker] Set list cells to a fixed width
With no fixed width, the list needs to calculate the maximum cell width,
so it needs to render all icons, but because to render them we also want
to know their position in the list we trigger the same call that started
this code-path, this means that we are looping trying to layout and
paint in a quadratic number of times.

Having a fixed width AND height avoids this process altogether, since
the List's layout manager knows everything it needs to layout the list
items.

Bug: 199811322
Test: manual, noticeable performance improvement for 100+ drawables
Change-Id: Ia0aa7891edff094e29fa4fb046bfc87671eb05a7