Update Studio lint plugin build process
This CL replaces the old mechanism for building the
Studio lint plugin with one based on the latest
recommendations; this is basically the
https://github.com/JetBrains/intellij-platform-plugin-template
project customized to work with our build setup.

This also updates to the latest versions of the
checks and makes it work with 2021.3 (and drops
support for 2020.3 and 2021.1).

Finally, it adds a test.

Test: Manual (new checks tested in the lint codebase)
Bug: None
Change-Id: Id6c69e9599ff6492ace68e465a23d79f85742d5e