Autoupdate of 3 submodules
*tools/adt/idea updated to e16d908a0ae139be20f26ee9429b323892d95560
Move GradleAndroidModel to project-system-gradle (Step 2)

(1) Move `GradleAndroidModel` and its dependencies to `project-system-gradle`;

(2) Do not move tests yet;

(3) Do not move any code which can be moved but which is not a dependency.

Bug: 147586218
Test: n/a
Change-Id: I450e1e2702cbe4fa690b99b63c9e20f0beb6ab90

*tools/studio/google/cloud/tools updated to 469be37ddeb0ab42b83cab9cba5619a2be1d67d3
Move GradleAndroidModel to project-system-gradle (Step 2)

Move `GradleAndroidModel` and its dependencies to `project-system-gradle`.

Bug: 147586218
Test: n/a
Change-Id: Ic8d2aea027d30ead236b55227a1dea74c91180b5

*tools/vendor/google updated to 1b1dd1aa12eb5312433c7b4b7eb2e139b3f40317
Move GradleAndroidModel to project-system-gradle (Step 2)

Move `GradleAndroidModel` and its dependencies to `project-system-gradle`.

Bug: 147586218
Test: n/a
Change-Id: Ib3477e670fe80e11c32bcd09f376443c638f6d72