Update lint fix diffs to label auto-applied fixes as such
Lint checks can be marked as autofixable, which means they will be
performed when the user runs the `lintFix` gradle target (or via `lint
--apply-suggestions`).

However, the autofix state was not included in the lint descriptions in
fix unit tests. This means that it's easy to miss whether a fix is or
isn't auto fixable.

This CL updates the description to label these fixes as auto fixable,
and updates existing tests to include the new descriptions.

Test: Existing (updated)
Bug: None
Change-Id: I5b3bfb5c420e06216d535f0aecc220420c376511