Autoupdate of 3 submodules
*tools/adt/idea updated to e2eddf3ae90f445db215ca3d5bb3d280db2bc855
Add existing @VisibleForTesting violations to baseline

Lint's annotation analysis now takes into account
constructor and classbody properties using default
usage sites, which means it picks up additional
violations.

This CL updates baselines to record these existing
problems.

Test: N/A
Bug: None
Change-Id: I29ea4b6adcdc7d9890202c7c994e13ba85ae768c

*tools/analytics-library updated to 34623f49159f1e545017b16d9013a3ba5db495a4
Add existing @VisibleForTesting violations to baseline

Lint's annotation analysis now takes into account
constructor and classbody properties using default
usage sites, which means it picks up additional
violations.

This CL updates baselines to record these existing
problems.

Test: N/A
Bug: None
Change-Id: Ie6cf27ba6f101eaacc1e2b3f9f9b8c14aaea8c9a

*tools/base updated to ff4cb0a79fedc6d2ae64df47210bc4d840d31960
Apply implicit Kotlin attributes from properties

If you specify a Kotlin annotation without a usage site on a
constructor property or a class property, that annotation will *only*
be applied to the parameter, property or field, it will never be
applied to the generated getters and setters for the property. This
means that the annotations are not included in analysis of usages of
the property getters and setters.

This CL updates lint to (when possible) search for these additional
properties and include them in analysis; currently, this is true
for properties analyzed as source, but not for properties referenced
in bytecode.

Fixes: 200048369
Fixes: 148905488
Test: New unit tests included
Change-Id: Id5c0dc4a6b4ad9e75d2a35d67246a2de227add6a