APK Sign: Fix passwords loading problem due to uncertain modality
Logging debug of repeated opening of the dialog showed how bug with
password missing can happen:
Passwords are initially scheduled to be retrieved when dialog
is constructed. After password retrieval we schedule setting it back
to the field in invokeLater call under modality of that field. And
here is the problem, nothing guarantees that dialog is already shown
and logging debug showed that sometimes this modality resolves to
NON_MODAL. In this case a runnable setting the password were not called
until dialog is closed leaving field empty.

In this particular case we do not care about modality, we just want to
set the password to the field in the UI thread, we are not changing any
models (as described on ModalityState javadoc).

This simple hotfix should help now and can be backported to C and B
branches, but we probably want to refactor dialog init process to
avoid such situation at all.

Bug: 205165814
Test: N/A
Change-Id: I53cdec2c5d8882221ca8d3feabf119fd8d8f11a7