Autoupdate of 2 submodules
*tools/adt/idea updated to 7419022a8d2242582d468c4699ed91282d936033
Add basic support for detector options

This CL updates a few places in the UI where we list lint
issues to include information about available options for
the issue.

Test: Unit tests included
Fixes: 207517590
Change-Id: I804f3f5b0ad5725bf7ddf0b2c159d3ac8d61aebf

*tools/base updated to c9cbcad6b3baaf3ddbd0549839ca3aa2cda8c892
Add basic support for detector options

This CL adds support for a simple "option" concept: an issue can
register one or more options, where the value type is for now limited
to simple value types like strings, boolean and numbers. By registering
the options and providing metadata like an explanation or for numeric
options, valid ranges, lint can include documentation for these in HTML
reports, issue documentation, explanation popups in the IDE, and so on.
It also gives a nicer way to look up options from detectors than the
previous mechanism which involved looking up values directly from the
configuration.

(It also fixes a bug in the TextFormat conversion from raw to text,
and updates the TextReporterTest to test the reporter in a better way.)

Test: Unit tests updated
Fixes: 207517590
Change-Id: I90ffdce5a13c97ed2864ce3a800af0b7851b2147