Refactor Preview cancellation flow
Change the cancellation flow to take advantage of the use of coroutines
and its cancellation. Instead of manually checking the progress
indicator status, this change creates a utility that allows a coroutine
to be launched to automatically follow the lifecycle of a progress
indicator.
If the progress indicator stops, the coroutine will also be stopped.
Reversely, if the coroutine is cancelled (or finishes), the progress
indicator will automatically be stopped.

As part of this change, also improve the refresh logging cycle to
attach a request id to trace all the progress. This is done via a Logger
decorator than can prefix the logs with any information.

Bug: N/A
Test: New tests added for launchWithProgress and LoggerWithFixedInfo
Change-Id: I20fed12892b81d194e7d8d9c3df63f05d320ee10