Use @Classpath path sensitivity for JdkImageInput to avoid cache misses
jrt-fs.jar is copied from jdk to create the JdkImageInput which is
the input of JavaCompileTask. jrt-fs.jar might only differ in
timestamps and that shouldn't be considered of forming a different
JdkImageInput. This CL makes sure JavaCompileTask gets a cache hit when
jrt-fs.jars differ in timestamp by using @Classpath because it ignores
JAR files names.

Bug: 209638107
Test: existing
Change-Id: I59cc104a9563a258b419c3f80db152a08b5ec35d