Prepare for not handling Blaze project system differently
in android.core resource repositories.

There is currently one place in the codebase where for the purpose of
modelling of Android resources, resource repositories process non-Gradle
project systems differently than the Gradle one. This is
`TestAppResourceRepository`, which calls
`AndroidModuleSystem.getAndroidLibraryDependencies(ANDROID_TEST)` only
in the case of the Gradle project system.

This change makes `getAndroidLibraryDependencies(ANDROID_TEST)` return
the empty list and ths allows to remove the special case.

Bug: 147586218
Test: n/a
Change-Id: I6fba800da8fcbd4a91ce806307dc01e25845c87a