Autoupdate of 2 submodules
*tools/adt/idea updated to e0b787fba60442bc8ce797595a8bde49332d890f
Remove myDebuggeeModuleProvider from AndroidDebuggerState

As we are trying untangle Debuggers from anything but Clients and debugger
setting I'm cleaning up AndroidDebuggerState.
AndroidDebuggerState should contain only debugger pure settings,
because we reuse this state between different modules/apps.
We still need facet, but I hope to remove it in future CLs.

The myDebuggeeModuleProvider is equal to the module in the facet,
 because facet received via AndroidFacet.getInstance(module).
See AndroidRunConfigurationBase.createLaunchTasksProvider.

Bug: 210440201
Test: refactoring
Change-Id: I0b5437cfbc714bc658e9c67c2d88520fb82f8a47

*tools/vendor/google updated to c8340d92622fc11e41f519d70b2b990f7b3f0de6
Remove myDebuggeeModuleProvider from AndroidDebuggerState

As we are trying untangle Debuggers from anything but Clients and debugger
setting I'm cleaning up AndroidDebuggerState.
AndroidDebuggerState should contain only debugger pure settings,
because we reuse this state between different modules/apps.
We still need facet, but I hope to remove it in future CLs.

The myDebuggeeModuleProvider is equal to the module in the facet,
 because facet received via AndroidFacet.getInstance(module).
See AndroidRunConfigurationBase.createLaunchTasksProvider

Bug: 210440201
Test: refactoring
Change-Id: I563f765a6dcb6e4328fd87132e325ab8b1fdad1e