Autoupdate of 2 submodules
*tools/adt/idea updated to 79f7e33b5871d8b4392fd787653c1409e969f7ab
211770054: Preserve quickfix order in the IDE

When there are multiple quickfixes registered, IntelliJ will show them
in *alphabetical* order, not the logical order they were registered in.

This CL attempts to work around that by using inspection priorities.
IntelliJ only has a concept of top/high/normal/low priorities, not a
general rank, so we cannot guarantee the order, but for short lists of
2-4 items it should be fine, and that's nearly always the case.

(Most of this fix updates various existing quickfixes to extend
DefaultLintQuickFix instead of implementing the LintQuickFix interface
directly which involves passing the display name to the super
constructor; then we add a priority property on DefaultLintQuickFix
which is used to remember the priority sorting for when we construct
IntentionActions from these.)

Test: AndroidLintInspectionBaseTest#testSortingPriority
Fixes: 211770054
Change-Id: I59d8990d1e1be903148711525fd79c6a9900d77f

*tools/base updated to 75a60ec7fe9d76a11261af96f5bacc0fcb37a8ea
211770054: Quickfix logical order is not used in the IDE

Fixes: 211770054
Test: LintFixVerifierTest and TypedefDetectorTest
Change-Id: Ida0fd8cd21d19177f9871206a0f6be74c0d1ecd1