Fix game tools broken from last IJ merge
Not sure what in the merge broke the tools but it looks like
ApplicationInfo now requires a plugin element, without this a
NPE is encountered on start up.
Changing the order of the included elements in the manifest.
Without this it seems like a duplicated plugin is loaded which
prevents the Java plugin from loading.

Test: manual
Bug: 202683062
Change-Id: I86cd6fdcac4b0bada15663142e5228ed63a47c8f