goldfish-opengl: add Linux meson build for Vulkan cereal
This adds a Meson build for Vulkan cereal.  Meson is well-liked
as an alternative to cmake, autogen, etc.

To build:

mkdir amd64-build/
meson amd64-build/
ninja -C amd64-build

This assumes the Vulkan headers are already installed in the
system.

The best bet is to define Vulkan headers internally in
goldfish-opengl to avoid the hassle of downloading them.

BUG=202552093
TEST=compile

Change-Id: Iab65f1cc6fde2ffcb1a07e65a44f2341cb9b84ef