[Dsl] Deeper copy when populating from subprojects
When we populate a subproject's initial properties from the contents of
a subprojects (or allprojects) block, we must be careful to copy
anything that might be mutated (for example, when adding individual
elements from a subproject's own configuration).

Define a new interface to retrieve a block's
PropertiesElementDescription, and use it when populating to create a
copy of the block.  Implement this interface for dependencies and
repositories, which have noticeable user-visible issues without this
correctness fix.  Subsequent CLs will implement and test this for
other Dsl blocks.

In order to maintain the correctness of the model's view of Dsl
elements in files, be sure to update elements' states from
APPLIED (effectively meaning "not present in this file") to
TO_BE_ADDED when modifying such an internal node.  (If nothing is
added to the element, it will be removed again as none of its leaves
will be in a non-APPLIED state).  Also be sure when removing an
element to remove it from all its holders, to preserve the existing
semantics.

(This CL addresses issue IDEA-285387)

Bug: 147139838
Test: added
Change-Id: Ie8dd1f5c3d764b73da14581d618a592f0285dc71