Autoupdate of 2 submodules
*tools/adt/idea updated to 8e8002f1d3427c0569b75a49024b76ccbf52c3e4
Move non-specific assemble() to GradleBuildInvoker

Assembling an IDE project should not require each consumer to find
the list of root modules to build. This is implementation details.

This is to reduce usages of `ProjectStructure.getLeafModules` to one
invocation only before it can safely replaced.

Also: Stop handling empty projects in `MakeGradleProjectAction`. We do
      not do it consistently in other actions.

Bug: n/a
Test: n/a
Change-Id: I5d080ee63e872f82751c78d98f03eabda3f4dc75

*tools/vendor/google updated to f0e9caaae535aad4f4a6c2a66f9359136640ced7
Move non-specific assemble() to GradleBuildInvoker

Assembling an IDE project should not require each consumer to find
the list of root modules to build. This is implementation details.

Bug: n/a
Test: n/a
Change-Id: I64457430692ed4cc189def6f37ac50037e6d3454