Remove strong references to ModuleClassLoader
Remove some unneeded strong references to the ModuleClassLoader to avoid
leaks. This includes removing the use of the disposer in some cases
since it creates strong references from the object tree.

IntelliJ also has some leaks where the focus is retained for already
disposed editors. This CL ensures that as much of the UI structure is
disposed in the WorkBench/DesignSurface when the editor is closed.

Bleak test fails but the problem does not reproduce manually by
verifying strong references via the profiler.

Bug: 213452854
Test: Existing tests pass. Bleak still fails.
Change-Id: Ic0f9626c0dbddc05d585d187b481536a48361408