Autoupdate of 2 submodules
*prebuilts/r8 updated to ed56ff453bd811de8f0646bed3622b3e8d2d668b
Update D8 R8 to 3.3.8-dev

This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.3.8-dev
Test: ./gradlew check
Bug:

Change-Id: I5804d7f1db70079f1e1e1cb0bca154ce5fdd6928

*tools/base updated to 5c6c85966810c452bb4263ba05cc68699f5c0ffd
Update test setup to pass program on classpath.

The update to applymapping in R8 requires mapped items to be on
classpath.

Bug:
Test: //tools/base/build-system/builder:tests
Change-Id: Id13a2e65c83dd13fd128490b4b2b1127c14fb517