Fix duplicated split editor controls for Kotlin/Java files
We use SeamlessTextEditorWithPreview to open Java/Kotlin files, and
manually control the visibility of the split editor toolbar to be able
to hide it when a file doesn't have custom views or Composables.

When IntelliJ introduced the floating controls, they changed the
split editor logic in a way that it hides/shows the toolbar if the
floating controls are enabled/disabled.

We're currently disabling the floating controls by setting a Registry
value, but if we enabled them, we we'd  end up with duplicate controls,
i.e. the fixed toolbar and floating controls.

This CL changes SeamlessTextEditorWithPreview to take into account the
floating controls when setting the split editor toolbar visibility.

Test: Manual
Bug: 214389534
Change-Id: I19a965643bc25dbf1bfa39a91efb067022548ae9