Autoupdate of 2 submodules
*tools/adt/idea updated to 449c14c3ab110ed96626bb28a761f02f7f03f479
Update BasicLayoutInspectorUITest

For context, this test was commented out during the migration of layout
inspector to app inspection framework because there was some difficulty
faking the transport / app inspection layer.

This change adapts the test to work with app inspection by faking the
transport layer (see parent change). It can connect to a fake process on
a fake device, and receive faked events, just like in unit tests.

However, most of the testing logic is still commented out because the
protocols have radically changed and will need to be updated by the LI
team.

Test: this is it.
Bug: 176097799
Change-Id: Iae56b2f0acbcafc636fc12d6fbc245dfe2ed6630

*tools/vendor/google updated to ea9397117a58f0c263a44e3d48affd8eaab2c773
Update transport's channelName getter

Test: N/A
Bug: 176097799
Change-Id: I9d2f9db3df552d52666d1ca215d1b611309bdf89