Autoupdate of 2 submodules
*tools/adt/idea updated to dbcbfa94df78871f33b16cd6c5713adef92a6b01
110692931: ObsoleteSdkInt should check RequiresApi annotations

Updates tests for fix applied to the lint codebase.

Test: This is it
Fixes: 110692931
Change-Id: I3f504e1705d1796e2634554d7487b4bfa707d98d

*tools/base updated to d535a88d4adb88248c693117ccca7020a16fb7a1
Allow annotation handler to visit annotations themselves

Until now, the annotation usage support in lint made it possible
to visit annotated elements -- references to classes and members
that were annotated, perhaps via an outer context or from an
override. However, it did not actually visit annotations themselves,
for example to enforce specific parameters on an annotation or
to check that they're used in the right context.

This CL adds a new usage type precisely for this purpose.

It then uses this in two places:

(1) In the ApiDetector, in order to flag obsolete usages of @RequiresApi
    (where @RequiresApi asks for an API that isn't higher than the
    effective minSdkVersion.). This fixes:
    110692931: ObsoleteSdkInt should check RequiresApi, TargetApi, and SdkSuppress annotations
    as well as the XML attribute targetApi (bug 37140910).

(2) In the IgnoreWithoutReasonDetector, replacing a manual element
    handler subclass with this more direct mechanism.

Test: Unit test included
Fixes: 110692931
Fixes: 74130329
Fixes: 37140910
Change-Id: I252056d210ed0b0ee59f19f3e09cb1f8383b1154