On partial fragment deopts, update return PC only when necessary
JITed code in debuggable apps don't have instrumentation stubs
installed and hence don't have an entry in the instrumentation
stack. In the case of a deopt the last frame that is popped
from instrumentation stack could be different from that of the
actual stack. Hence we shouldn't update the return PC
unconditionally incase of a partial fragment deopt.

Test: art/test.py -t 989-method-trace-throw
Bug: 214001306
Change-Id: I10cf74b3a1481429502b46e9d69b5274cce0f000