Autoupdate of 2 submodules
*tools/adt/idea updated to fe92e0b6c064b7d4b91d4cc32c75ad725ee19351
Store permissions bits as Int rather than decimal String

Bug: 205005362
Test: updated
Change-Id: I60d77813d72e0c4332e28d61e352aeb26005c664

*tools/base updated to 11c766b4f7ce343726786acf8e783071d0eddf16
Check permissions in SyncCommandHandler

In order to test error handling in Device Explorer, we need to simulate
permissions checking in the sync service.

Also, store permissions as Int (rather than decimal String) in
DeviceFileState to make it easier to work with.

Bug: 205005362
Test: updated
Change-Id: If507effb36aaeb232548b575bda02897083b0e36