Don't try to acquire a read lock if there's a pending write action
This is to avoid the deadlock that a lengthy background read action
is never interrupted by a coming write action request, and such
pending write action results in blocking the other thread to acquire
the read lock, which blocks back the thread holding the read lock.

For example, thread A starts a read action, which spawns a future
task and busy waiting the result in thread B which needs a read
lock. Here comes a write request on EDT which is blocking thread
B to get the read lock, and this write action is blocked because
thread A doesn't abort the execution and release the read lock
even it's supposed to respond to such write request.

Ideally, thread A should abort the read execution and notify the
future task spawned by itself to cancel the work, as it doesn't
make sense to proceed.

However, this read lock is never released because
1) there's not enough invocations of `ProgressManager#checkCanceled()`
2) we don't have a handler or pass a progress indicator to support
cancellation on the spawned future task.

So here's a tactical fix that we check
1) if the read access is allowed and 2) if there's a pending write
action. If the read access is allowed, we just proceed the work.
If there's a pending write action but no read access allowed yet,
we yield to write request and cancel the current process.

Please note, it's safer to cancel the process when this particular
condition is met, instead of just relying on checking the canceled
process. As cancellation happens after `isWriteActionPending` bit
is flipped, which means there's chance we miss this write request.

Bug: 215201912
Test: manual
Change-Id: I2424b90ee839d1de1b3ff4206ef6027df3501707