Protect against concurrent modification during disposal
This change prevents occasion Studio crash due to a race condition in the native code.

The VideoDecoder class is no longer Disposable. Most of the code was moved from VideoDecoder into an inner class, DecodingContext that is allocated at the beginning of the VideoDecoder.start method and closed at the end. The cleanup code that was previously located in VideoDecoder.dispose has been moved to DecodingContext.close.

Test: N/A
Bug: N/A
Change-Id: I2a25bffcb053ba0f2e985abc82e89b36f9792dd8