Autoupdate of 2 submodules
*prebuilts/studio/sdk/darwin updated to 5a4f483b3cefba05e7f62ceb8caaa060538bfa18
Copy android-31 sources into android-32

The source files in this change are not real source files. They are
copied from android-31 so that tests that rely on the latest platform
will pass.

Bug: 210480197
Test: N/A
Change-Id: I57ac55bfc28207cea527633bfbcd70dc39f3275e

*prebuilts/studio/sdk/windows updated to 2766ea5fedbdb0856427952231bda54a07cc22a3
Copy android-31 sources into android-32

The source files in this change are not real source files. They are
copied from android-31 so that tests that rely on the latest platform
will pass.

Bug: 210480197
Test: N/A
Change-Id: I7221b5a698003774c004abc214648e01994d1736