Autoupdate of 4 submodules
*tools/data-binding updated to e7560ee1bc63b39534d0e06c5dd8f0bc7c180566
Add repeatOnLifecycle to data binding

This replaces the previous implementation that used
`LifecycleCoroutineScope.launchWhenCreated`. The
`repeatOnLifecycle` API is the recommended way to collect
flows from the UI since it restarts the coroutine when the
lifecycle moves in and out of a target state instead of
suspending it, as `launchWhenCreated` does.

Fixes: b/184935697
Test: Manual testing and KotlinTestApp integration tests
Change-Id: I464558942243559cf19e0164d30626239a69be88

*prebuilts/tools updated to 9c918a7c4802403cf3cefbb998b5a11df7245135
Add androidx lifecycle 2.4.0 dependencies

Change-Id: I41489896878d33e56cb91c18df583ec1de38121e

*tools/adt/idea updated to 2fcb8826c2e552a3595590cef180f8332bec6354
Update androidx lifecycle dependencies to 2.4

Change-Id: I4cffe13b5a792efcf7f6cbaad4b85399732b3ea7

*tools/base updated to 9c487423799a3f5fe50d17271ad0000974ed3906
Update androidx lifecycle dependencies to 2.4

Change-Id: Ib4dad38aad2d535252079eb6bb8d8a2b907781ec