Autoupdate of 2 submodules
*tools/adt/idea updated to 7706efe408d909203d09aa40fdb7f9e33ea142c7
Add the comments to note for the themes for material3

To note that the corresponding files need to be modified when the
contents of the themes are modified.

Also fixes the TemplateTest.testNewBasicActivityMaterial3 to correctly
generate the project as a material3 project.

Bug: 209907794
Test: Manually followed the steps at the bug
Change-Id: Icec8e36387148395fdf140fc22e031fc08e8db2e

*tools/base updated to 40461000df8a963ccf4f3aeecd066ce12546377a
Generate a set of material3 themes when creating a new Activity

The set of themes for the BasicActivity (Material3) are created
when creating a new project.

But when creating a new Activity under a flavor, the corresponding
themes are not created.

This CL creates the set of themes every time a new Activity is created
(when a new project is created, the themes in the Activity recipe are
ignored)

Bug: 209907794
Test: Manually followed the steps at the bug
Change-Id: I1731b62fb77432e71a75fb517773dac13606dba4