Increase scroll speed
On windows the previous scroll speed was very slow, and now it's faster
(though still significantly slower than in text editors).
On linux it was slow before, but not as slow as on windows. Now the
scroll speed is similar to what it is in text editors.
Mac uses it's own mechanism for this, so the fix has no impact there
(but also the problem doesn't exist there).

Test: none
Fixes: 112158835
Change-Id: I304242dcec41d4551342ee2fa415182711ca77d5