Check cancellation in mock file upload
Since testFileSystemTree_ContextMenu_Upload_SingleFile_Cancellation_Works
uses the mock file system for the upload, we need to check for
cancellation there. Otherwise, it is schedule-dependent whether the test
passes: if the cancellation occurs quickly, DeviceExplorerController will
see it and cancel, but if the mock file transfer is reached, cancellation
will be ignored and the test will timeout.

Fixes: 214591919
Test: fixed
Change-Id: I4a8d1cd26742c1588fb2bdc16fd1845df912f25a