[ResPicker] Fix resources not updating
Adding resources within the ResourcePickerDialog may cause it to not
update automatically. Since the ResourceNotificationManager waits with
invokeLater to notify listeners, which are blocked in a modal dialogue.

This fixes it by waiting for the repository to update under a modal
dialogue whenever the picker dialog regains focus.

This way we can update the picker without worrying about jankiness from
interfering with user input.

Fixes: 209567318
Test: updated ui test, unit test to proof implemented logic
Change-Id: I3c32970780173ebb65e90cb6c853f2208d19bff8