Autoupdate of 2 submodules
*tools/adt/idea updated to 1a6d2c8dc8b087a1f1ccfaee56923fc688c80945
Factor out function to create NDK release name

This is so that it can also be used in tests to adjust
baseline snapshots to not have fixed NDK versions

Bug: NA
Test: NA
Change-Id: I4f9a760c251ca350efbdfce027d3b569181fa04d

*tools/vendor/google updated to 5104dd696486cdef372c2673c172ec9bdd688b05
Unfix more NDK baselines

In these baseline tests, replaces the "NDK r23b (...)" text with
equivalent but without hard-coded version number. This is to make it
easier to upgrade NDKs in the future.

Bug: NA
Test: NA
Change-Id: Ibc15ac56de9f14f0e4bf1252a9c037354e7e6f12