Autoupdate of 2 submodules
*tools/adt/idea updated to f03a1e3c61bc9cb35e4940a8848811d7ad750fb0
Update GradleModuleSystemIntegrationTest with profileable android:enabled

ag/16717873 adds a new ManifestSystemProperty enum value.
GradleModuleSystemIntegrationTest asserts the number of enum values, so
it needs to be updated.

Bug: 207395929
Test: GradleModuleSystemIntegrationTest
Change-Id: I2d9631307f36822eb1d552ecf624d1b6b1a27d76

*tools/base updated to 4bee4975de49154e269dc4fe9643b01eaad08c43
Append android:enabled to profileable

Explicitly Sets the profileable attribute `android:enabled` to true when
the profiling mode is set to `profileable`, in the event the attribute
is overriden to false elsewhere.

Bug: 207395929
Test: ProfileableTest
Change-Id: I80a013381a332bcd21036b15d07bc24727a4f165