Fix event re-selection
Equality weren't defined on `CaptureNodeAnalysisModel`, so each time
we set it in the multi-selection model, it was considered a different
selection that triggered the heavier-weight event that caused a
re-population of the tab bar and giving focus to the most recent tab.

This CL converts `CaptureNodeAnalysisModel` to Kotlin while we're at
it. By declaring the class `data`, Kotlin generates `equal`/`hash`
using structural equality.

Fixes: 217786325
Test: add new
Change-Id: I4a873f45caf48928a44f867e9319e72d9aa34ad8