Autoupdate of 3 submodules
*prebuilts/tools updated to 00d43e4c271d2ec6993a227833a6e173f3fb238b
Add dependencies for AGP 7.0 and 7.1

To enable adding of tests for these versions to Android Studio.

Bug: 216916629
Test: See I25f3127b5f7c2a4a5d67e4e7625042b2ab83b1c6
Change-Id: I650cc569b902049b08ab765dc50639fade3ce207

*tools/adt/idea updated to 35b753764cf1c282f0a7af1cbd403c4675e40fbb
Use corresponding lint version

Bug: N/A
Test: Included
Change-Id: I0ed9caa4cb679c00e823a5172ba8e85d411dd58c

*tools/base updated to 5532780d3657f7242e58e7f958431756b159f742
Add dependencies for AGP 7.0 and 7.1

To enable adding of tests for these versions to Android Studio.

This is the lint and aapt2 artifacts that are loaded dynamically and are
required for a successful sync of some test projects.

This change only adds artifacts, rather than replacing older artifacts
as would be done for compilation dependencies. This is to enable some
targeted sync tests to verify project import from each supported version
of the Android Gradle Plugin. (When we decide to increase the minimum
supported version in Android Studio the tests can be updated and the
older prebuilts can then be removed.)

Bug: 216916629
Test: See I25f3127b5f7c2a4a5d67e4e7625042b2ab83b1c6
Change-Id: Ia89e717243b054b58422489d8db49439d8a53e28