Fix initial UI freeze when loading large perfetto trace
This change moves the heavy work to the background to eliminate a UI
freeze of several seconds when loading a large trace. It also factors
out the common pattern that was at least used in loading the heap dump.

The simple implementation of `runAsync` in FakeIdeServices also reveals
a test bug from quickly selecting different heap dumps, which this
change also fixes.

Fixes: 217583421
Bug: 172345706
Bug: 217583905
Bug: 217804316
Test: add new
Change-Id: I18abc6653462577ccdbd97c50913183a15184424