Add forceCompileSdkPreview to AAR metadata file
When an Android library module is compiled against a preview SDK (e.g.,
Tiramisu), then AGP will add that information to the AAR metadata file
(e.g., adding the line `forceCompileSdkPreview=Tiramisu`).

In a subsequent change, AGP will enforce the rule that any module
consuming such a library must be compiled against an SDK with the exact
same preview codename.

The reason to enforce such a rule is because if an app uses a library
compiled against a preview SDK and is uploaded to the play store, the
app might crash on devices running the finalized API (if there are
changes between preview and final), but devs are prevented from
uploading an app to the play store if compiling against a preview SDK.

Bug: 218900545
Test: AarMetadataTaskTest (unit and integration tests)
Change-Id: I81f72bd402614d1e294584f67400e46c8dafe9f4