218525937: Revamp AIDL support
The AIDL editor support was really out of date, missing a lot of
language features added to AIDL over the years:

* Annotations
* Enums (Android 11+)
* Unions (Android 12+)
* Nested type declarations (Android T+)

In this CL, these language features have been added to the parser, and
(perhaps even more importantly) the grammar has been updated to support
many other corner cases that the previous grammar did not (e.g. u8
suffix in numbers, missing keywords like cpp_header and const, support
for constant initializer expressions). It also removes the keywords
"flattenable" and "rpc" which are not valid in AIDL.

I tried to adjust the old grammar to add these features but quickly
gave up and instead wrote a new grammar, which tries to closely
resemble the authoritative lex and yacc grammars in the AIDL tool
source code in the platform. (The grammar is current as of AOSP CL
048af158).

I did purposely change the names in some cases; this is noted in the
grammar file, and was done to ensure that the generated PSI elements
follow some canonical names (because there isn't a way to rename the
PSI classes generated from the BNF rule names.)

I also deliberately separated out the difference between a declared
name and a name reference (see below).

In addition to the grammar updates, I made the following improvements
to the AIDL support at the same time:

* Added support for go-to-declaration on references (for example, if
  the IDL file has a List<File> type for a parameter, you can ctrl
  click on the List or the File tokens and jump to the corresponding
  classes). AIDL has a handful of "builtin" type mappings which it will
  also use (before searching the imported symbols).

* Added parenthesis and bracket matching (when typing, and highlighting
  matching for symbol next to caret)

* Updated the annotator such that it highlights different semantic
  types in the parsed tree: annotations, function names, constants, and
  type references.

* Added more tests (there were no tests for the annotator). Moved the
  AIDL tests from the android/ plugin into the android-lang/ plugin
  (they appear to have been accidentally left in the Android plugin
  when the rest of the android-lang support was pulled out?). Added
  additional tests for the new language features and made the find
  usages test be more comprehensive (list usages, check resolving, etc.)

  Added many more parsing tests (lifted from the AIDL codebase). Added
  toString() overrides for most of the AIDL PSI elements such that the
  pretty-printed trees in highlighting tests are easier to read and
  mentally verify.

* I moved the generated .flex file into the gen/ folder; it's derived
  from the .bnf file and should not be edited manually. I didn't
  realize this at first and was adjusting the lexical analyzer first
  before looking at the parser, only to realize I had wasted my time
  when I moved on to fixing the parser. Having it in gen/ makes a lot
  more sense.

Overall this improves the AIDL support such that we correctly handle
the current language features, but there's a lot of room for
improvement left -- code completion, refactoring support (e.g.
participate in change signature refactorings), formatting and
indentation handling, etc.

Bug: 218525937
Test: Many unit tests included
Change-Id: Icc87b4cd8b7639c1a08c8e030a8262b3a6948bd4