Use TestData to determine if a task is a test
This is a temporary workaround for Android Studio.
The long-term plan is for Jetbrains to fix it:
https://youtrack.jetbrains.com/issue/IDEA-288709

Jetbrains changed the way they configure a test run configuration and
determine if they should render test output in the test-output window.

They now rely on the ExternalProject's tasks to be available and to
check against them in order to determine if a given task (from a run
config) should be run as a test.

This change broke test rendering on Android Studio because we disabled
fetching the Gradle tasks per default due to Sync performance issues.

Since we also inject TestData into the module nodes, containing
information about what tests each module has, and the corresponding test
task, we can also use this information to tell what is a test task.

TestData is injected by all the other plugins that use Gradle to run
tests, so we can get TestData for non-Android modules as well.

Bug: 217700611
Test: Verified manually
Change-Id: I508541480b12b0a2635993b46b22626dad33ce89