Autoupdate of 2 submodules
*tools/adt/idea updated to bbe9c52015e1365351782fc387e1f1fd555148be
Use main modules in Android run configurations

This is to allow things that try to get information from the module in
the run configuration to have a none empty module. This is
also the more "correct" setup as we are building from the sources of
this main module.

We also remove patching of run configurations from pre-qualified name
versions since with this we can no longer guess the module of the
runconfiguration. This was also broken in general after MPSS since we
run over all module based configuration whichs will set any none android
configuration to the empty holder module.

We also fix a small issue with Android test run configurations being
wrongly marked as invalid by adjusting the accepted modules for these
configurations.

Fixes: 217847811
Test: snapshots
Change-Id: Ib68d3f5e9b0a6740f111789900d41448c6804302

*tools/vendor/google updated to e8dc1a1b47c5dc580f05c0d44ee3055753ec9932
Update snapshots for run configuration module change

We are changing the Android run configurations to be based upon the main
module that contains most of the information about the module. Currently
they are based on holder modules which results in some issues fetching
information from them.

Bug: N/A
Test: snapshots updated
Change-Id: I50aaec5f996d6c2c2c84b57a7db3a165ecc03634